requirements.txt
asgiref==3.5.2
audiomath==1.16.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.19.5
botocore==1.22.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cryptography==35.0.0
distlib==0.3.6
Django==4.1.2
docopt==0.6.2
filelock==3.8.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Js2Py==0.71
MarkupSafe==2.0.1
numpy==1.21.4
packaging==21.3
paramiko==2.8.0
passlib==1.7.4
Pillow==8.3.1
pipi==1.0.1
pipwin==0.5.1
platformdirs==2.5.2
psutil==5.8.0
psycopg2==2.9.5
PyAudio @ file:///C:/Users/ASUS/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
pyjsparser==2.7.1
PyNaCl==1.4.0
pyparsing==3.0.6
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
qrcode==7.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SoundCard==0.4.1
sounddevice==0.4.3
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.29
sqlparse==0.4.3
tzdata==2022.5
tzlocal==4.1
urllib3==1.26.7
vboxapi==1.0
virtualenv==20.16.5
Werkzeug==2.0.1
wikipedia==1.4.0
